Refactor so anywhere controls derive from the native framework control
Refactor to avoid the need for a wrapper object for anywhere controls.
Here's the new scheme:

HostFrameworkAnywhereControl is the parent class for AnywhereControl.
When building, you build against a stub version of this class.
At runtime, there's a different version of the class that's used to power the
UI, specific to the host UI Framework. In this way, an AnywhereControl on WPF
ends up deriving from System.Windows.Controls.Control , like any other WPF
control. Avoiding the wrapper makes things for efficient (one less allocation per
control) and makes the architecture (in some ways) simpler,
with anywhere controls acting more like native platform controls

When there's a need to mix different UI frameworks together, an AssemblyLoadContext
can be used for that so that each framework gets the appropriate version of
HostFrameworkAnywhereControl (future thing).