Remove ReadLocalValue
Avalonia and MAUI don't have good support (apparently) for ReadLocalValue.
And it's not clear that we actually need it - that anyone will use that APIs.
For both reasons, remove that API for now - we'll add it back in the future
if/when there's a clear need for it, keeping things simple until there's some clear
need to add complexity.